+++ src/core/class-store.php
@@ -3,7 +3,7 @@
  * Store
  *
  * @author Takuto Yanagida
- * @version 2024-03-22
+ * @version 2024-03-26
  */
 
 namespace nt;
@@ -17,12 +17,49 @@ require_once( __DIR__ . '/class-query.php' );
 
 class Store {
 
+	/**
+	 * URL.
+	 *
+	 * @var string
+	 */
 	protected $_dirUrl;
+
+	/**
+	 * Root directory.
+	 *
+	 * @var string
+	 */
 	protected $_dirRoot;
+
+	/**
+	 * Configuration.
+	 *
+	 * @var array<string, mixed>
+	 */
 	protected $_conf;
+
+	/**
+	 * Types.
+	 *
+	 * @var Type
+	 */
 	protected $_type;
+
+	/**
+	 * Taxonomies.
+	 *
+	 * @var Taxonomy
+	 */
 	protected $_taxonomy;
 
+	/**
+	 * Constructor for the class.
+	 *
+	 * @param string $ntUrl   The URL of the directory.
+	 * @param string $ntDir   The root directory.
+	 * @param string $dataDir The data directory.
+	 * @param array<string, mixed> $conf The configuration for the constructor.
+	 */
 	public function __construct( string $ntUrl, string $ntDir, string $dataDir, array $conf ) {
 		$this->_dirUrl  = $ntUrl;
 		$this->_dirRoot = $ntDir;
@@ -32,23 +69,56 @@ class Store {
 		$this->_taxonomy = new Taxonomy( $dataDir, $conf );
 	}
 
-	public function type()    : Type     { return $this->_type; }
+	/**
+	 * Gets the type of the post.
+	 *
+	 * @return Type The type of the post.
+	 */
+	public function type(): Type { return $this->_type; }
+
+	/**
+	 * Gets the taxonomy of the post.
+	 *
+	 * @return Taxonomy The taxonomy of the post.
+	 */
 	public function taxonomy(): Taxonomy { return $this->_taxonomy; }
 
 
 	// ------------------------------------------------------------------------
 
 
+	/**
+	 * Gets the URL of the post.
+	 *
+	 * @param string  $id      The ID of the post.
+	 * @param ?string $subPath The subpath of the post.
+	 * @return string The URL of the post.
+	 */
 	public function getPostUrl( string $id, ?string $subPath ): string {
 		if ( $subPath === null ) $subPath = $this->getSubPath( $id );
 		return $this->_dirUrl . $subPath . $id . '/';
 	}
 
+	/**
+	 * Gets the directory of the post.
+	 *
+	 * @param string  $id      The ID of the post.
+	 * @param ?string $subPath The subpath of the post.
+	 * @return string The directory of the post.
+	 */
 	public function getPostDir( string $id, ?string $subPath ): string {
 		if ( $subPath === null ) $subPath = $this->getSubPath( $id );
 		return $this->_dirRoot . $subPath . $id . '/';
 	}
 
+	/**
+	 * Creates the archive and sub-path of the post.
+	 *
+	 * @param string $type            The type of the post.
+	 * @param string $dateRaw         The raw date of the post.
+	 * @param bool   $ensureExistence Whether to ensure the existence of the directory.
+	 * @return string[] The directory and sub-path of the post.
+	 */
 	public function createArchAndSubPath( string $type, string $dateRaw, bool $ensureExistence = false ): array {
 		$ret = $this->_conf['archive_by_type'] ? "$type/" : 'post/';
 		if ( $this->_conf['archive_by_year'] ) {
@@ -62,6 +132,12 @@ class Store {
 		return [ $this->_dirRoot . $ret, $ret ];
 	}
 
+	/**
+	 * Gets the subpath of the post.
+	 *
+	 * @param string $id The ID of the post.
+	 * @return ?string The subpath of the post.
+	 */
 	public function getSubPath( string $id ): ?string {
 		$ds = $this->_getSubPaths();
 		foreach ( $ds as $d ) {
@@ -70,6 +146,12 @@ class Store {
 		return null;
 	}
 
+	/**
+	 * Gets the sub-paths of the post.
+	 *
+	 * @param ?string $type The type of the post.
+	 * @return string[] The sub-paths of the post.
+	 */
 	private function _getSubPaths( ?string $type = null ): array {
 		$typeDirs = $type ? [ "$type/" ] : $this->_type->getTypeDirAll();
 		if ( ! $this->_conf['archive_by_year'] ) return $typeDirs;
@@ -78,14 +160,16 @@ class Store {
 		foreach ( $typeDirs as $t ) {
 			if ( ! is_dir( $this->_dirRoot . $t ) ) continue;
 			$dir = dir( $this->_dirRoot . $t );
-			while ( false !== ( $fn = $dir->read() ) ) {
-				if ( $fn[0] === '.' || $fn[0] === '_' || $fn[0] === '-' ) continue;
-				if ( preg_match( '/[^0-9]/', $fn ) ) continue;
-				if ( strlen( $fn ) !== 4 ) continue;
-				if ( is_file( $t . $fn ) ) continue;
-				$ret[] = $t . $fn . '/';
+			if ( $dir instanceof \Directory ) {
+				while ( false !== ( $fn = $dir->read() ) ) {
+					if ( $fn[0] === '.' || $fn[0] === '_' || $fn[0] === '-' ) continue;
+					if ( preg_match( '/[^0-9]/', $fn ) ) continue;
+					if ( strlen( $fn ) !== 4 ) continue;
+					if ( is_file( $t . $fn ) ) continue;
+					$ret[] = $t . $fn . '/';
+				}
+				$dir->close();
 			}
-			$dir->close();
 		}
 		$ret = array_merge( $ret, $typeDirs );  // For option compatibility
 		return $ret;
@@ -95,13 +179,28 @@ class Store {
 	// ------------------------------------------------------------------------
 
 
+	/**
+	 * Gets the post.
+	 *
+	 * @param ?string $id The ID of the post.
+	 * @return ?Post The post.
+	 */
 	public function getPost( ?string $id ): ?Post {
-		if ( $id === null ) return null;
-		$p = new Post( $id, $this->getSubPath( $id ) );
+		if ( null === $id ) return null;
+		$sp = $this->getSubPath( $id );
+		if ( null === $sp ) return null;
+		$p = new Post( $id, $sp );
 		if ( ! $p->load() ) return null;
 		return $p;
 	}
 
+	/**
+	 * Gets the post with next and previous posts.
+	 *
+	 * @param string|null          $id   The ID of the post.
+	 * @param array<string, mixed> $args The arguments.
+	 * @return array{Post|null, Post|null, Post|null}|null The post with next and previous posts.
+	 */
 	public function getPostWithNextAndPrevious( ?string $id, array $args = [] ): ?array {
 		if ( $id === null ) return null;
 		$posts = $this->_getPosts( $args );
@@ -120,6 +219,12 @@ class Store {
 		return [ $prev, $posts[ $idIndex ], $next ];
 	}
 
+	/**
+	 * Gets the posts.
+	 *
+	 * @param array<string, mixed> $args The arguments.
+	 * @return array<string, mixed> The posts.
+	 */
 	public function getPosts( array $args = [] ): array {
 		$page = $args['page'] ?? 1;
 
@@ -138,6 +243,13 @@ class Store {
 		return ['posts' => $ret, 'size' => $size, 'page' => $pageIdx + 1, 'page_count' => ceil( $size / $perPage ) ];
 	}
 
+	/**
+	 * Gets the count by date.
+	 *
+	 * @param string               $type The type of the post.
+	 * @param array<string, mixed> $args The arguments.
+	 * @return array<string, mixed>[] The count by date.
+	 */
 	public function getCountByDate( string $type, array $args ): array {
 		$args += [ 'status' => Post::STATUS_PUBLISH ];
 		$paths = $this->_getPostTypeSubDirs( $args );
@@ -169,6 +281,12 @@ class Store {
 	// -------------------------------------------------------------------------
 
 
+	/**
+	 * Gets the posts.
+	 *
+	 * @param array<string, mixed> $args The arguments.
+	 * @return Post[] The posts.
+	 */
 	private function _getPosts( array $args ): array {
 		$args += [ 'status' => Post::STATUS_PUBLISH ];
 		$posts = [];
@@ -181,6 +299,13 @@ class Store {
 		return $posts;
 	}
 
+	/**
+	 * Loads the matched posts.
+	 *
+	 * @param string                 $root   The root directory.
+	 * @param array<string, mixed>   $args   The arguments.
+	 * @param array<string, mixed>[] &$posts The posts.
+	 */
 	private function _loadMatchedPostAll( string $root, array $args, array &$posts = [] ): void {
 		$paths = $this->_getPostTypeSubDirs( $args );
 		$ret = [];
@@ -192,6 +317,12 @@ class Store {
 		}
 	}
 
+	/**
+	 * Gets the post type subdirectories.
+	 *
+	 * @param array<string, mixed> $args The arguments.
+	 * @return string[] The post type subdirectories.
+	 */
 	private function _getPostTypeSubDirs( array $args ): array {
 		if ( $this->_conf['archive_by_type'] ) {
 			if ( empty( $args['type'] ) ) {
@@ -205,6 +336,14 @@ class Store {
 		return ['post/'];
 	}
 
+	/**
+	 * Loads the matched information.
+	 *
+	 * @param string                 $root  The root directory.
+	 * @param string[]               $paths The paths.
+	 * @param array<string, mixed>   $args  The arguments.
+	 * @param array<string, mixed>[] &$ret  The return value.
+	 */
 	private function _loadMatchedInfoAll( string $root, array $paths, array $args, array &$ret = [] ): void {
 		$is_trash = ( isset( $args['status'] ) && $args['status'] === '_trash' );
 		if ( $is_trash ) unset( $args['status'] );
@@ -212,10 +351,24 @@ class Store {
 		$this->_loadMatchedInfoAllInternal( $root, $paths, $query, $is_trash, $ret );
 	}
 
+	/**
+	 * Loads the matched information internally.
+	 *
+	 * @param string                 $root    The root directory.
+	 * @param string[]               $paths   The paths.
+	 * @param Query                  $query   The query.
+	 * @param bool                   $isTrash Whether the post is in trash.
+	 * @param array<string, mixed>[] &$ret    The return value.
+	 */
 	private function _loadMatchedInfoAllInternal( string $root, array $paths, Query $query, bool $isTrash, array &$ret = [] ): void {
 		foreach ( $paths as $path ) {
-			if ( ! is_dir( $root . $path ) ) continue;
+			if ( ! is_dir( $root . $path ) ) {
+				continue;
+			}
 			$dir = dir( $root . $path );
+			if ( ! ( $dir instanceof \Directory ) ) {
+				continue;
+			}
 			while ( false !== ( $fn = $dir->read() ) ) {
 				if ( $fn[0] === '.' || $fn[0] === '_' ) continue;
 				if ( ! $isTrash && $fn[0] === '-' ) continue;
@@ -235,6 +388,13 @@ class Store {
 		}
 	}
 
+	/**
+	 * Loads the information.
+	 *
+	 * @param string $postDir The directory of the post.
+	 * @param string $pid     The ID of the post.
+	 * @return ?array<string, mixed> The information.
+	 */
 	private function _loadInfo( string $postDir, string $pid ): ?array {
 		$infoPath = $postDir . '/' . Post::INFO_FILE_NAME;
 		$json     = false;
@@ -256,6 +416,12 @@ class Store {
 	// ------------------------------------------------------------------------
 
 
+	/**
+	 * Creates a new post.
+	 *
+	 * @param string $type The type of the post.
+	 * @return ?Post The new post.
+	 */
 	public function createNewPost( string $type = 'post' ): ?Post {
 		$dateRaw = date( 'YmdHis' );
 		list( $archPath, $subPath ) = $this->createArchAndSubPath( $type, $dateRaw, true );
@@ -279,6 +445,13 @@ class Store {
 		return null;
 	}
 
+	/**
+	 * Ensures the uniqueness of the post ID.
+	 *
+	 * @param string $archPath The path of the archive.
+	 * @param string $dateRaw  The raw date.
+	 * @return ?string The unique post ID.
+	 */
 	private function _ensureUniquePostId( string $archPath, string $dateRaw ): ?string {
 		$id = $dateRaw;
 		if ( ! $this->_checkIdExists( $archPath, $id ) ) return $id;
@@ -289,6 +462,13 @@ class Store {
 		return null;
 	}
 
+	/**
+	 * Checks if the ID exists.
+	 *
+	 * @param string $archPath The path of the archive.
+	 * @param string $id       The ID.
+	 * @return bool Whether the ID exists.
+	 */
 	private function _checkIdExists( string $archPath, string $id ): bool {
 		return
 			file_exists( $archPath       . $id ) ||
@@ -300,6 +480,12 @@ class Store {
 	// ------------------------------------------------------------------------
 
 
+	/**
+	 * Writes the post.
+	 *
+	 * @param Post $post The post.
+	 * @return Post The post.
+	 */
 	public function writePost( Post $post ): Post {
 		$post->save();
 		$id = $post->getId();
@@ -312,6 +498,11 @@ class Store {
 		return $post;
 	}
 
+	/**
+	 * Removes the post.
+	 *
+	 * @param string $id The ID of the post.
+	 */
 	public function removePost( string $id ): void {
 		if ( $id[0] === '-' ) {
 			$pd = $this->getPostDir( $id, null );
@@ -336,6 +527,11 @@ class Store {
 		}
 	}
 
+	/**
+	 * Restores the post.
+	 *
+	 * @param string $removed_id The ID of the removed post.
+	 */
 	public function restorePost( string $removed_id ): void {
 		if ( $removed_id[0] !== '-' ) return;
 		$subPath = $this->getSubPath( $removed_id );
@@ -352,6 +548,11 @@ class Store {
 		}
 	}
 
+	/**
+	 * Empties the trash.
+	 *
+	 * @param string $type The type of the post.
+	 */
 	public function emptyTrash( string $type ): void {
 		$ps = $this->_getPosts( [ 'type' => $type, 'status' => '_trash' ] );
 		foreach ( $ps as $p ) {
@@ -360,6 +561,11 @@ class Store {
 		}
 	}
 
+	/**
+	 * Empties the temporary directories.
+	 *
+	 * @param string $type The type of the post.
+	 */
 	public function emptyTemporaryDirectories( string $type ): void {
 		global $nt_session;
 		$temps = $nt_session->listTemporaryDirectories();
@@ -367,8 +573,14 @@ class Store {
 		$ds = $this->_getSubPaths( $type );
 		foreach ( $ds as $d ) {
 			$dir =  $this->_dirRoot . $d;
-			if ( ! is_dir( $dir ) ) continue;
-			foreach ( scandir( $dir ) as $fn ) {
+			if ( ! is_dir( $dir ) ) {
+				continue;
+			}
+			$fns = scandir( $dir );
+			if ( false === $fns ) {
+				continue;
+			}
+			foreach ( $fns as $fn ) {
 				if ( $fn !== '.' && $fn !== '..' && $fn[0] === '_' ) {
 					$t = "$dir$fn/";
 					if ( is_dir( $t ) && ! in_array( $t, $temps, true ) ) {
@@ -379,18 +591,26 @@ class Store {
 		}
 	}
 
+	/**
+	 * Deletes all files and directories in the specified directory.
+	 *
+	 * @param string $dir The directory to delete.
+	 */
 	static public function deleteAll( string $dir ): void {
 		$dir = rtrim( $dir, '/' );
 		if ( ! is_dir( $dir ) ) {
 			Logger::error( __METHOD__, 'The post directory does not exist', $dir );
 			return;
 		}
-		foreach ( scandir( $dir ) as $fn ) {
-			if ( $fn !== '.' && $fn !== '..' ) {
-				if ( is_dir( "$dir/$fn" ) ) {
-					self::deleteAll( "$dir/$fn" );
-				} else {
-					unlink( "$dir/$fn" );
+		$fns = scandir( $dir );
+		if ( is_array( $fns ) ) {
+			foreach ( $fns as $fn ) {
+				if ( $fn !== '.' && $fn !== '..' ) {
+					if ( is_dir( "$dir/$fn" ) ) {
+						self::deleteAll( "$dir/$fn" );
+					} else {
+						unlink( "$dir/$fn" );
+					}
 				}
 			}
 		}
